Cleaned up some dependencies on the way to removing the cast.py file
