+++ keanu-python/keanu/vertex/base.py
@@ -1,4 +1,3 @@
-import collections
 from typing import List, Tuple, Iterator, Union, SupportsRound, Optional
 from typing import cast as typing_cast
 
@@ -10,8 +9,8 @@ import keanu as kn
 from keanu.base import JavaObjectWrapper
 from keanu.context import KeanuContext
 from keanu.tensor import Tensor
-from keanu.vartypes import (tensor_arg_types, wrapped_java_types, shape_types, numpy_types, runtime_tensor_arg_types,
-                            runtime_primitive_types, runtime_wrapped_java_types)
+from keanu.vartypes import (tensor_arg_types, wrapped_java_types, shape_types, numpy_types, runtime_wrapped_java_types,
+                            runtime_primitive_types, runtime_numpy_types, runtime_pandas_types)
 
 k = KeanuContext()
 
@@ -170,16 +169,9 @@ class Vertex(JavaObjectWrapper, SupportsRound['Vertex']):
 
     @staticmethod
     def __parse_arg(arg: Union[vertex_constructor_param_types, shape_types]) -> JavaObject:
-        print("PARSING ARG: " + str(arg))
-#        if isinstance(arg, runtime_tensor_arg_types):
-#            print("Making new Constant")
-#            return kn.vertex.const.Const(arg).unwrap()
         if isinstance(arg, runtime_wrapped_java_types):
             print("Unwrapping Arg")
             return arg.unwrap()
-#        elif isinstance(arg, collections.Collection) and all(isinstance(x, runtime_primitive_types) for x in arg):
-#            print("Converting to java long array")
-#            return k.to_java_long_array(arg)
         else:
             raise ValueError("Can't parse generic argument. Was given {}".format(type(arg)))
 
@@ -199,19 +191,39 @@ class Vertex(JavaObjectWrapper, SupportsRound['Vertex']):
 class Double(Vertex):
 
     def cast(self, v: tensor_arg_types) -> tensor_arg_types:
-        from keanu.cast import cast_tensor_arg_to_double
         return cast_tensor_arg_to_double(v)
 
 
 class Integer(Vertex):
 
     def cast(self, v: tensor_arg_types) -> tensor_arg_types:
-        from keanu.cast import cast_tensor_arg_to_integer
         return cast_tensor_arg_to_integer(v)
 
 
 class Boolean(Vertex):
 
     def cast(self, v: tensor_arg_types) -> tensor_arg_types:
-        from keanu.cast import cast_tensor_arg_to_bool
         return cast_tensor_arg_to_bool(v)
+
+
+def __cast_to(arg: tensor_arg_types, cast_to_type: type) -> tensor_arg_types:
+    if isinstance(arg, runtime_primitive_types):
+        return cast_to_type(arg)
+    elif isinstance(arg, runtime_numpy_types):
+        return arg.astype(cast_to_type)
+    elif isinstance(arg, runtime_pandas_types):
+        return arg.values.astype(cast_to_type)
+    else:
+        raise TypeError("Cannot cast {} to {}".format(type(arg), cast_to_type))
+
+
+def cast_tensor_arg_to_double(arg: tensor_arg_types) -> tensor_arg_types:
+    return __cast_to(arg, float)
+
+
+def cast_tensor_arg_to_integer(arg: tensor_arg_types) -> tensor_arg_types:
+    return __cast_to(arg, int)
+
+
+def cast_tensor_arg_to_bool(arg: tensor_arg_types) -> tensor_arg_types:
+    return __cast_to(arg, bool)
